requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.7